requirements.txt
langchain>=0.1.0
openai==1.7.2
ipykernel
pymupdf>=1.23.0
arxiv==2.1.0
nltk==3.8.1
tiktoken==0.5.2
chromadb==0.3.29
pypdf==3.17.4
lark==1.1.9
streamlit==1.30.0
watchdog==3.0.0
hydra-core>=1.3.0